requirements.txt
aiogram==3.13.1
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
